:recycle: refactor: Utility funcs are really mapping funcs
